Move the commands to the Encryption namespace
